### 2021-08-26 #126 @whedon
[PRE REVIEW]: Virtual training on virtual environments: an online open-source introduction to conda
**Submitting author:** @marisalim (<a href="http://orcid.org/0000-0003-2097-8818">Marisa Lim</a>)
**Repository:** <a href="https://github.com/nih-cfde/training-and-engagement" target ="_blank">https://github.com/nih-cfde/training-and-engagement</a>
**Version:** n/a
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
