### 2021-08-31 #126 @allisonhorst
Hi @shanamatthews, thank you for signing up to review for the Journal of Open Source Education! Would you like to contribute a review for this submission, "Virtual training on virtual environments: an online open-source introduction to conda"? Please let me know if you're available and willing to review this submission. Thank you!
