### 2021-09-01 #126 @allisonhorst
Thank you @colbrydi, we really appreciate your willingness to review for JOSE! No worries at all about learning the review process. I am new to it, too, and here to help (with the support of the JOSE editorial board). I'll go ahead and add you as a reviewer for this paper and kick off the review process. Please reach out any time with questions, and I'll follow up with some information within the next two days.
