### 2021-09-01 #126 @shanamatthews
> Hi @shanamatthews, thank you for signing up to review for the Journal of Open Source Education! Would you like to contribute a review for this submission, "Virtual training on virtual environments: an online open-source introduction to conda"? Please let me know if you're available and willing to review this submission. Thank you!

Hi @allisonhorst! I’m happy to review as well.
