### 2021-09-01 #130 @whedon
[REVIEW]: Virtual training on virtual environments: an online open-source introduction to conda
**Submitting author:** @marisalim (<a href="http://orcid.org/0000-0003-2097-8818">Marisa Lim</a>)
**Repository:** <a href="https://github.com/nih-cfde/training-and-engagement" target ="_blank">https://github.com/nih-cfde/training-and-engagement</a>
**Version:** v2021.12

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".
