### 2021-09-02 #130 @shanamatthews
Hey @allisonhorst - it looks like there are some broken links in the instructions:
- [JOSE conflict of interest policy](https://github.com/openjournals/jose/blob/master/COI.md)
- The code of conduct link on this page: https://jose.theoj.org/about#code_of_conduct (https://github.com/openjournals/jose/blob/main/CODE_OF_CONDUCT.md)

Possibly these files need to be copied over from the joss repository?
