### 2021-09-02 #130 @allisonhorst
Thanks for the heads up @shanamatthews!

@labarba can you update these links for JOSE reviewer COI and CoC (or let me know who at JOSE I should reach out to)? Thanks!
