### 2021-09-06 #130 @labarba
Yikes! It is the same as the [JOSS COI policy](https://github.com/openjournals/joss/blob/master/COI.md). We will investigate that bug in our issue generator (JOSE is a fork of JOSS and we recently reorganized the repositories).
