### 2021-09-08 eff37f5
AWS JOSE Paper
