+++ paper/paper.md
@@ -1,0 +1,120 @@
+---
+title: 'A virtual training module for introducing the use of Amazon Web Services'
+tags:
+  - reproducible research
+  - AWS
+authors:
+  - name: Abhijna Parigi^[co-first author]
+    orcid: 0000-0003-2097-8818
+    affiliation: 1  
+  - name: Marisa C.W. Lim^[co-first author]
+    orcid: 0000-0002-5861-3737
+    affiliation: 1
+  - name: Saranya Canchi^[co-first author]
+    orcid: 0000-0002-9404-7162
+    affiliation: 1
+  - name: Jose Sanchez
+    affiliation: 1
+  - name: Jeremy Walter
+    orcid: 0000-0001-7074-9760
+    affiliation: 1
+  - name: Amanda Charbonneau
+    orcid: 0000-0001-7376-7702
+    affiliation: 1
+  - name: C. Titus Brown
+    orcid: 0000-0001-6001-2677
+    affiliation: 1
+affiliations:
+  - name: Lab for Data Intensive Biology, School of Veterinary Medicine, University of California Davis
+    index: 1  
+date: 8 September 2021
+bibliography: paper.bib
+---
+
+
+# AWS Jose Paper
+
+# Summary
+
+We present our lesson material and resources for introducing the use of Amazon Web Services (AWS, https://aws.amazon.com/) for cloud computation. This lesson was developed for the Common Fund Data Ecosystem (CFDE), an NIH initiative that aims to promote data re-use and cloud computing for biomedical research. The lesson materials, technology set-up instructions, and our instructional experiences can serve as a resource for prospective instructors who wish to re-use and remix our materials for teaching AWS and cloud computing.
+
+# Statement of need
+
+Amazon's Elastic Compute Cloud (EC2) is one of the most popular cloud-based platforms for computing and storage. Yet, there are very few freely available and comprehensive resources specifically designed for beginners. To fill this gap, we developed a curriculum to teach novice users (with minimal command-line and no cloud computation experience) about cloud computing in general, and AWS in particular.
+
+We aimed to create modular content that requires minimal setup or operating system (OS) dependence, in line with the short-format (2-hour) instructor-led virtual workshops run by the CFDE. To this end, our lesson incorporates the use of IAM (Identity and Access Management) credentials that allow all workshop participants to log in to a common AWS 'root' account. The use of IAM credentials for teaching purposes, in combination with an OS-agnostic shell interface that runs in any web browser, greatly minimizes the initial setup time and allows participants to receive a comprehensive, free-of-charge introduction to working with AWS and the cloud.
+
+All our tutorials are hosted on the [CFDE training website](https://training.nih-cfde.org/en/latest/) and include short instructional videos to accompany the self-guided text tutorials. If a user is unable to attend one of our live workshops, they can pick another medium of learning and work through content at their own pace.
+
+# Curriculum
+
+## Learning Objectives
+
+As discussed in our [online lesson](https://training.nih-cfde.org/en/latest/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws1.html), after completing our workshop, a learner should be able to:
+- Set up an AWS cloud computer
+- Connect to the AWS cloud computer
+- Run multiple screen sessions and switch between them
+- Download data from an AWS cloud computer to their local computer
+- Terminate an AWS cloud computer
+
+## Content
+
+In the live workshops, we begin with a brief introduction to the basics of cloud computing and AWS, discuss commonly used (AWS) terminologies, and highlight the advantages of using AWS (over non-cloud platforms).
+
+Next, we lead participants through launching and connecting to AWS instances. Here, participants log in to a common workshop AWS account using the IAM user credentials that we provide them.
+
[…]
