+++ paper/paper.md
@@ -1,9 +1,9 @@
 ---
 title: 'A virtual training module for introducing the use of Amazon Web Services'
 tags:
-  - reproducible research
   - AWS
   - cloud computing
+  - reproducible research
 authors:
   - name: Abhijna Parigi^[co-first author]
     orcid: 0000-0003-2097-8818
