### 2021-09-09 #130 @shanamatthews
Hello! I was able to spend some time reviewing this morning and found that the materials met most requirements and overall seemed high quality - great work!

I've pasted in the parts of the review checklist where I saw gaps. I was not able to find documentation in the repo or in the paper to help other users adopt or adapt the module, although there is plenty of information about how the original group of authors have used it. Perhaps a short section on "teaching with this module" would be appropriate in both the repo and the paper?

> ### Documentation
> * [x]  **A statement of need:** Do the authors clearly state the need for this module and who the target audience is?
> * [x]  **Installation instructions:** Is there a clearly stated list of dependencies?
> * [ ]  **Usage:** Does the documentation explain how someone would adopt the module, and include examples of how to use it?
> * [x]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support
>
> ### JOSE paper
> * [x]  **Authors:** Does the `paper.md` file include a list of authors with their affiliations?
> * [x]  **A statement of need:** Does the paper clearly state the need for this module and who the target audience is?
> * [x]  **Description:** Does the paper describe the learning materials and sequence?
> * [ ]  Does it describe how it has been used in the classroom or other settings, and how someone might adopt it?
> * [x]  Could someone else teach with this module, given the right expertise?
> * [x]  Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?
> * [x]  **References:** Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
