### 2021-09-09 #130 @marisalim
Hi @shanamatthews - thank you for reviewing our submission! Do you think additional information on [this wiki page for instructors](https://github.com/nih-cfde/training-and-engagement/wiki/Conda-workshop-teaching-notes) would work? We were trying to stay within the word limit for the paper, but are happy to add more information for using the module in the wiki!
