### 2021-09-15 nih-cfde/training-and-engagement#472 @colbrydi
Consider adding --from-history option to "conda list --export"
I suggest that you consider including the --from-history argument when advising to export the conda list.  Although it will not reproduce the exact environment I have found that the argument makes for much easier to understand yml files.  Just a recommendation, I realize you also don't want to overburden the learner with too many details.

BTW. Thank you for this great tutorial.  I happen to be working on a similar one for python and having this as a reference will save me a lot of time!  I learned a lot.
