### 2021-09-15 nih-cfde/training-and-engagement#473 @colbrydi
Consider adding some narrtion to the conda videos
Although, I realize that all of the key information is in the accessible text of the tutorial and the video is only there as additional reference, I still recommend adding verbal content to the videos and allow YouTube to generate CC transcripts. I found the video confusing without the text and it took me a second to find the text describing what was in the video.  I realize this is a minor point but if the video just had someone narrating what was going on I think it could add a lot.  Overall great tutorial, --thank you
