### 2021-09-15 #130 @colbrydi
Hello I completed my review today.  I happen to also be working on a  tutorial for using conda environments in Python using jupyter notebooks and this could not have come at a better time for me.   I learned a lot from your submission and look forward to being able to use and reference it in my own work.  I submitted a couple of issue but they are just recommendations and I feel this is ready for publication. -- Thank you!
