### 2021-09-15 nih-cfde/training-and-engagement#472 @ctb
thanks, Dirk!

I've used it a few times and it never did much for me, IIRC, but I think maybe I wasn't using it properly. We should try it again and include it if it slims down the list at all!
