### 2021-09-20 #130 @shanamatthews
I also think this submission is ready for publication. I've signed off on all the review criteria, but would love to see more explicit info on how external folks can adapt the training in the wiki page, as mentioned.
