### 2021-10-01 #130 @labarba
Hi everybody! 👋

I see that this submission targets a short tutorial on virtual environments and conda that is a slice of a repository hosting lots of other teaching materials:
https://github.com/nih-cfde/training-and-engagement/tree/dev/docs/General-Tools/Introduction-to-Conda

In addition, the same group is submitting for review another slice of the collection of materials, targeting an introduction to AWS services:
https://github.com/openjournals/jose-reviews/issues/132

This poses a few questions: first, it kind of looks like salami-slicing an educational initiative into a bunch of thin papers. Second, the JOSE publication workflow assumes one repository per paper, which gets archived in Zenodo after review, linking the DOI of the archive with the DOI of the paper. In this case, the large repository contains a lot of material not associated with the paper, and/or several JOSE papers would link to the same GitHub repo and Zenodo archive.

To move forward, in my opinion, the authors would have to re-organize their materials into substantial chunks that justify a scholarly publication, each in a separate GitHub repository that can have its own Zenodo archive in the end.

I suggest this submission be withdrawn in order to address this. Feel free to comment here with your thoughts or other ideas.
