### 2021-10-01 #130 @ctb
hi @labarba,

> first, it kind of looks like salami-slicing an educational initiative into a bunch of thin papers.

I suppose it could look like that, but we've offered these workshops separately and assessed them separately, which is reflected in the papers. They can be reused and offered separately, and are on (quite) different topics.

Also, maybe I am mistaken, but we were already pressed for space given the length restrictions (and reviewers asked for more). I'm not quite sure how we can do the lessons justice if we combine them.

Last but not least, the reviewers seem OK with the content. It is strange to have this come up now, after the review.

(This is an ongoing initiative for developing open educational materials that are delivered frequently and (in at least some cases) will be updated regularly, and that's how we're submitting them.)

>Second, the JOSE publication workflow assumes one repository per paper, which gets archived in Zenodo after review, linking the DOI of the archive with the DOI of the paper. In this case, the large repository contains a lot of material not associated with the paper, and/or several JOSE papers would link to the same GitHub repo and Zenodo archive.

OK, but note that this reorganizes our work in a way that doesn't match how they were developed or are delivered, which seems pretty artificial. I understand it matches the JOSE model and so we can do this reorganization. But it would probably be good to note this requirement for organization explicitly & up front in the submission instructions.

Let us know what you all think and how to proceed - thanks!
