### 2021-10-01 #130 @labarba
I noticed the situation upon seeing the new submission come in
https://github.com/openjournals/jose-reviews/issues/132
...and realizing it had the same target repository as this one.

A 2-hr workshop is on the small side for what we view as a "minimum publishable unit," comparing the scholarly effort to a research paper. We want to be sure that a JOSE paper is respected in the authors' track record and CV, and thus aim for more weighty teaching modules.

I'm raising with the editorial board the issue of multiple papers having a common target repository. We've not encountered this situation before. In our publication model the repository archive DOI gets linked to the paper DOI via the Crossref metadata.
