### 2021-10-04 #130 @labarba
After talking with friends over at JOSS, I found that we discussed at length the "big code repo / paper covering a submodule" situation, and we untangled it so it's a non-issue. Here is the conversation  about that: https://github.com/openjournals/joss/issues/443

**Highlights**:
- different papers that present sub-modules or ring-fenced new functionality for review (that's for JOSS), where the source code is all in one big repository, are not a problem
- the archive DOIs will be different, because they are deposited at a different moment in time, so the linked DOIs in the Crossref metadata work as designed
- the request that archive metadata (title and author list) match the paper is relaxed in this case, as it doesn't work well
- multiple `paper.md` files in the same repo can be handled by working from a branch during paper review (in fact, this situation was what motivated teaching `whedon` to compile from branch)

---

Regarding the size of what constitutes substantial scholarly effort for a JOSE publication, that is an ongoing conversation and I'm calling an editorial board meeting where we will discuss it. (The boundary will always be fuzzy, but we need to improve our guidance to authors regardless.)
