### 2021-10-18 f3b2974
add sept and oct workshops
