### 2021-11-01 2b01eef
update found resource helpful stat
include stats from latest workshop
