+++ paper/paper.md
@@ -69,7 +69,7 @@ We asked participants to fill out short pre- and post-workshop surveys to help i
 
 To address the challenge of gauging learning and participation in remote workshops, we employed several successful teaching strategies: checking in with yes/no questions and exercises (responses via chat or Zoom raised hand reaction), using a standardized computing interface with workshop commands script, and providing a help breakout room. With an instructional team of 4-5, we recommend workshop sizes of 20-30 participants, as these strategies are less effective for lower instructor to participant ratios. 
 
-The survey feedback helped to refine material by reducing information overload that obscured learning objectives and to improve pacing. For example, we ask participants to launch binder environments upon arrival, so they are ready by the time we reach the hands-on portion. In addition, we now send [general resources for attendees](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees) before the workshop, along with specific [workshop notes](https://github.com/nih-cfde/training-and-engagement/wiki/Introduction-to-Conda) that are streamlined from our website lesson to guide instructors and participants. Our surveys show that ~82% of conda workshop participants read and found the general resources useful (n=39 responses). Participants also suggested adding time for more exercises and intermediate topics, highlighting the efficacy of hands-on workshops to help new users learn introductory material and develop confidence in exploring conda functionality further. 
+The survey feedback helped to refine material by reducing information overload that obscured learning objectives and to improve pacing. For example, we ask participants to launch binder environments upon arrival, so they are ready by the time we reach the hands-on portion. In addition, we now send [general resources for attendees](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees) before the workshop, along with specific [workshop notes](https://github.com/nih-cfde/training-and-engagement/wiki/Introduction-to-Conda) that are streamlined from our website lesson to guide instructors and participants. Our surveys show that 86.8% of conda workshop participants read and found the general resources useful (n=53 responses). Participants also suggested adding time for more exercises and intermediate topics, highlighting the efficacy of hands-on workshops to help new users learn introductory material and develop confidence in exploring conda functionality further. 
 
 In summary, these remote workshops have been largely successful, as participants reported that the pacing was not too slow or fast (Figure 1A) and a larger percentage of respondents agreed they would like to use conda for future projects in post- compared to pre-workshop surveys (Figure 1B).
 
