### 2021-11-05 #130 @allisonhorst
Hi all:

I feel the paper sufficiently addresses more rigorous expectations for short workshops discussed at the meeting. Including:

- The workshop has been run 4x with 94 total participants

- Authors provide helpful information on workshop instructional design, including structure of the teaching team (e.g. “The workshop is led by two co-instructors who both teach; a moderator who tracks questions, learner participation, and time; and additional helpers who offer one-on-one help in breakout rooms”), and a platform that works (on Zoom w/ Pangeo binders + RStudio)

- Authors make recommendations for workshop size (20 - 30) based on their experience

- Authors describe and report results from surveys used to gauge participant experience & pace

- Authors share links to general resources for attendees and workshop notes/additional resources. The general resources for attendees will save future teachers a substantial amount of time and frustration (e.g. helping workshop participants get set up before the workshop). The instructor notes are also useful for considering common points of confusion.

In summary, I think this submission does satisfy the more rigorous expectations for a short 2-hour workshop based on the last JOSE meeting.
