### 2021-11-05 #130 @labarba
Should some of the aspects listed be reflected better in the paper, @allisonhorst ?
