### 2021-11-15 cfb5ba9
clarify vidlet topics, add export option
per suggestions from Dirk's JOSE review
