### 2021-11-15 nih-cfde/training-and-engagement#473 @marisalim
thanks for this suggestion! adding more context for the vidlets on PR #475
