### 2021-11-19 #130 @marisalim
Thank you for the feedback @allisonhorst, @shanamatthews, and @colbrydi!
- Per @shanamatthews's suggestions, I've updated our wiki space to provide more information for preparing to use our materials for a workshop [here](https://github.com/nih-cfde/training-and-engagement/wiki/Conda-workshop-teaching-notes), along with a [template](https://hackmd.io/iEOJg-U6QvujQTzw4wpzRg) for workshop notes.
- Per @colbrydi's suggestions ([here](https://github.com/nih-cfde/training-and-engagement/issues/473) and [here](https://github.com/nih-cfde/training-and-engagement/issues/472)), I added more context for the website videos and added a note about the `--from-history` export conda option to the website ([merged PR](https://github.com/nih-cfde/training-and-engagement/pull/475)) - it's on our development branch right now, and will be merged into the public website at the next release date (we can move this up as needed).
- We taught one more Conda workshop in the time since the paper was submitted, so I've updated the participants and survey statistics information in the [paper.md](https://github.com/nih-cfde/training-and-engagement/blob/JOSE-conda-paper/paper/paper.md).

Please let me know if additional edits are needed!
