### 2021-11-19 ed773b0
Updates to AWS tutorial with new screenshots and correction to AMI choice
