+++ mkdocs.yml
@@ -29,7 +29,7 @@ plugins:
 
 markdown_extensions:
 - admonition
-#- def_list
+- def_list
 - pymdownx.critic
 - pymdownx.details
 - pymdownx.highlight
