### 2021-11-21 #130 @allisonhorst
Thank you for the update @marisalim !

I appreciate your work to address the reviewers' feedback, including updates to the teaching notes and the hackmd template for workshop notes, which I think are very helpful. A couple of minor suggestions:

1. Consider updating the paper text where you refer to the [workshop teaching notes](https://github.com/nih-cfde/training-and-engagement/wiki/Conda-workshop-teaching-notes#before-workshop), as it is not clear from the paper that those notes provide step-by-step recommendations for using and adapting your materials. Specifically, "For workshop preparation, we created resources for instructors, including teaching notes and steps for creating an RStudio binder" should be updated to clarify that these notes provide clear guidance on steps teachers can follow to reuse these materials for their own workshop, and should be the starting point to workshop setup. I think the paper undersells the teaching notes resource, and the language pointing to it can be strengthened to better describe the contents.

2. I think it would be helpful if the paper contained a concise list of the resources teachers will use when implementing this workshop, something like:

"The resources that teachers adopting this workshop will use are:

- Teaching notes with guidance for setting up & adapting the workshop
- hackmd template to create course-specific notes
- Existing website with workshop materials

..." (or similar)

I feel that having all of those resources listed, linked and briefly described all together in a single place in the paper would be helpful for a teacher looking to use the materials.

Please let me know if you have questions on this feedback or other aspects of the paper.
