### 2021-11-23 54d76b2
add resource list suggestions from allison
