+++ paper/paper.md
@@ -35,7 +35,7 @@ bibliography: paper.bib
 
 We present our lesson material and resources for teaching how to use conda (https://conda.io), a tool that streamlines software installation and version management using isolated environments, while providing multiple methods for reproducing and sharing software set ups. This training material was developed for the NIH Common Fund Data Ecosystem (CFDE), whose primary goal is to teach biologists computational tools that help make their analysis workflows FAIR: Findable, Accessible, Interoperable, and Reusable [@Wilkinson_etal_2016]. 
 
-We teach conda because it can be used in any research field that requires software installation, works on multiple computer operating systems (Mac, Windows, Linux), and does not require admin privileges. We piloted this material in March 2021 and have now taught the 2-hour online workshop four times to a diverse audience. The lesson material and accompanying resources provide information for instructors to adopt our hands-on workshop format and for learners to view past workshop notes or walk-through the lesson at their own pace from our [training website](https://training.nih-cfde.org/en/latest/General-Tools/Introduction-to-Conda/index.html).
+We teach conda because it can be used in any research field that requires software installation, works on multiple computer operating systems (Mac, Windows, Linux), and does not require admin privileges. We piloted this material in March 2021 and have now taught the 2-hour online workshop four times to a diverse audience. The lesson material and accompanying resources provide information for instructors to adopt our hands-on workshop format and for learners to view past workshop notes or walk-through the lesson at their own pace from our [training website](https://training.nih-cfde.org/en/latest/General-Tools/Introduction-to-Conda/).
 
 # Statement of Need
 
@@ -59,7 +59,13 @@ The lesson was designed to be self-contained and flexible: learners can go at th
 
 ## Workshop Instructional Design
 
-We designed the lesson structure to be a 2-hour remote course. The workshop is led by two co-instructors who both teach; a moderator who tracks questions, learner participation, and time; and additional helpers who offer one-on-one help in breakout rooms. For workshop preparation, we created resources for instructors, including [teaching notes](https://github.com/nih-cfde/training-and-engagement/wiki/Conda-workshop-teaching-notes) and [steps for creating an RStudio binder](https://github.com/nih-cfde/training-and-engagement/wiki/Making-an-RStudio-binder).
+We designed the lesson structure to be a 2-hour remote course. The workshop is led by two co-instructors who both teach; a moderator who tracks questions, learner participation, and time; and additional helpers who offer one-on-one help in breakout rooms. As a starting point for workshop preparation, we created several resources to provide clear guidance that teachers can follow to reuse these materials for their own workshops, including:
+- the [CFDE training website Intro to conda lesson](https://training.nih-cfde.org/en/latest/General-Tools/Introduction-to-Conda/) for self-guided learning about lesson content
+- [teaching notes](https://github.com/nih-cfde/training-and-engagement/wiki/Conda-workshop-teaching-notes) for before, during, and after workshop preparation using CFDE materials
+- a [workshop notes template](https://hackmd.io/iEOJg-U6QvujQTzw4wpzRg) to edit and share with workshop participants
+- the [CFDE Intro to conda workshop notes archive](https://github.com/nih-cfde/training-and-engagement/wiki/Introduction-to-Conda) for reference about past events
+- a [conda Q&A](https://github.com/nih-cfde/training-and-engagement/discussions/categories/conda-q-a) page for questions that have been asked during our workshops
+- [steps for creating an RStudio binder](https://github.com/nih-cfde/training-and-engagement/wiki/Making-an-RStudio-binder) to edit or create a new one
 
[…]
