### 2021-11-23 dd34c30
minor text edits
