+++ paper/paper.md
@@ -35,13 +35,13 @@ bibliography: paper.bib
 
 We present our lesson material and resources for teaching how to use conda (https://conda.io), a tool that streamlines software installation and version management using isolated environments, while providing multiple methods for reproducing and sharing software set ups. This training material was developed for the NIH Common Fund Data Ecosystem (CFDE), whose primary goal is to teach biologists computational tools that help make their analysis workflows FAIR: Findable, Accessible, Interoperable, and Reusable [@Wilkinson_etal_2016]. 
 
-We teach conda because it can be used in any research field that requires software installation, works on multiple computer operating systems (Mac, Windows, Linux), and does not require admin privileges. We piloted this material in March 2021 and have now taught the 2-hour online workshop four times to a diverse audience. The lesson material and accompanying resources provide information for instructors to adopt our hands-on workshop format and for learners to view past workshop notes or walk-through the lesson at their own pace from our [training website](https://training.nih-cfde.org/en/latest/General-Tools/Introduction-to-Conda/).
+We teach conda because it can be used in any research field that requires software installation, works on multiple computer operating systems (Mac, Windows, Linux), and does not require administrator privileges to configure. We piloted this material in March 2021 and have now taught the 2-hour online workshop five times to a diverse audience. The lesson material and accompanying resources provide information for instructors to adopt our hands-on workshop format and for learners to view past workshop notes or walk-through the lesson at their own pace from our [training website](https://training.nih-cfde.org/en/latest/General-Tools/Introduction-to-Conda/).
 
 # Statement of Need
 
 As a consequence of switching to fully remote-compatible training due to the COVID-19 pandemic, we designed online material for self-paced and virtual workshop learning modes. Existing conda tutorials typically teach learners how to run commands with the assumption that learners' computer configuration and software versions match those in the lesson, however software and system set-up can be time-consuming, and distract or dissuade learners from wanting to use conda.
 
-For our lesson, we specifically designed our materials to run on a binder cloud computing environment that is free and easy to launch [@binder]. This structure allows learners to focus on the course content because the computing environment provides standardized interface and software versions; works for Mac, Windows, and Linux users; and reduces remote troubleshooting challenges for learners and instructors. To follow this lesson, learners only need a stable internet connection and  web browser (i.e., Firefox, Chrome, Safari). While our lesson does include steps to install and set up conda on learners' own computers, it is taught after hands-on experience with conda, which allows them to decide whether it meets their research needs.
+For our lesson, we specifically designed our materials to run on a binder cloud computing environment that is free and easy to launch [@binder]. This structure allows learners to focus on the course content because the computing environment provides standardized interface and software versions; works for Mac, Windows, and Linux users; and reduces remote troubleshooting challenges for learners and instructors. To follow this lesson, learners only need a stable internet connection and  web browser (e.g., Firefox, Chrome, Safari). While our lesson does include steps to install and set up conda on learners' own computers, it is taught after hands-on experience with conda, which allows them to decide whether it meets their research needs.
 
 # Curriculum
 
@@ -59,7 +59,7 @@ The lesson was designed to be self-contained and flexible: learners can go at th
 
 ## Workshop Instructional Design
 
-We designed the lesson structure to be a 2-hour remote course. The workshop is led by two co-instructors who both teach; a moderator who tracks questions, learner participation, and time; and additional helpers who offer one-on-one help in breakout rooms. As a starting point for workshop preparation, we created several resources to provide clear guidance that teachers can follow to reuse these materials for their own workshops, including:
[…]
