### 2021-11-23 #130 @marisalim
Thanks for those suggestions @allisonhorst! I have edited the [paper](https://github.com/nih-cfde/training-and-engagement/blob/JOSE-conda-paper/paper/paper.md).
