### 2021-11-30 601e09f
add november workshops
