### 2021-11-30 b66cc99
fix conda workshop link
