+++ docs/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws5_Screen.md
@@ -2,7 +2,7 @@
 
 Screen or GNU Screen is a terminal multiplexer. You can start a terminal session and then open multiple screens inside that session. Processes running in Screen will continue to run when their window is not visible, even if your session gets disconnected/times out - you can even close the terminal/connection to AWS instance and commands running on screens will continue!
 
-The goal of this tutorial is to learn how to run and switch between multiple screen sessions. Check out the [screen cheat sheet](../../General-Tools/Cheat-Sheets/screen_cheatsheet.md) for commonly used commands.
+The goal of this tutorial is to learn how to run and switch between multiple screen sessions. Check out the [screen cheat sheet](../../General-Tools/Cheat-Sheets/screen_cheatsheet.md){:target="_blank"} for commonly used commands.
 
 ## Video Walkthrough
 <iframe id="kaltura_player" src="https://cdnapisec.kaltura.com/p/1770401/sp/177040100/embedIframeJs/uiconf_id/29032722/partner_id/1770401?iframeembed=true&playerId=kaltura_player&entry_id=1_g03komqj&flashvars[mediaProtocol]=rtmp&amp;flashvars[streamerType]=rtmp&amp;flashvars[streamerUrl]=rtmp://www.kaltura.com:1935&amp;flashvars[rtmpFlavors]=1&amp;flashvars[localizationCode]=en&amp;flashvars[leadWithHTML5]=true&amp;flashvars[sideBarContainer.plugin]=true&amp;flashvars[sideBarContainer.position]=left&amp;flashvars[sideBarContainer.clickToClose]=true&amp;flashvars[chapters.plugin]=true&amp;flashvars[chapters.layout]=vertical&amp;flashvars[chapters.thumbnailRotator]=false&amp;flashvars[streamSelector.plugin]=true&amp;flashvars[EmbedPlayer.SpinnerTarget]=videoHolder&amp;flashvars[dualScreen.plugin]=true&amp;flashvars[Kaltura.addCrossoriginToIframe]=true&amp;&wid=1_gu82i69s" width="608" height="402" allowfullscreen webkitallowfullscreen mozAllowFullScreen allow="autoplay *; fullscreen *; encrypted-media *" sandbox="allow-forms allow-same-origin allow-scripts allow-top-navigation allow-pointer-lock allow-popups allow-modals allow-orientation-lock allow-popups-to-escape-sandbox allow-presentation allow-top-navigation-by-user-activation" frameborder="0" title="Kaltura Player"></iframe>
@@ -201,4 +201,4 @@ If no other screen sessions are open, you will fall back to the original SSH ter
     ```
     Typing `exit` too many times will exit the entire terminal!
 
-Don't forget to check out our [screen cheat sheet](../../General-Tools/Cheat-Sheets/screen_cheatsheet.md) if you need a quick reference to screen commands.
+Don't forget to check out our [screen cheat sheet](../../General-Tools/Cheat-Sheets/screen_cheatsheet.md){:target="_blank"} if you need a quick reference to screen commands.
