+++ docs/Cloud-Platforms/Introduction_to_Amazon_Web_Services/introtoaws3.md
@@ -9,7 +9,7 @@ Follow along with these steps and/or watch our [walk-through tutorial](./introto
 
 ## Step 1: Log in to an AWS account
 
-* Go to [Amazon Web Services](https://aws.amazon.com)
+* Go to [Amazon Web Services](https://aws.amazon.com){:target="_blank"}
 * Click on <span class="highlight_txt">My Account</span>
 * Select  <span class="highlight_txt">AWS Management Console</span> from the drop down menu.
 * Alternatively, click on <span class="highlight_txt">Sign In to the Console</span>.
@@ -19,7 +19,7 @@ Follow along with these steps and/or watch our [walk-through tutorial](./introto
 
 !!! info "Account Setup"
 
-          If you need to create an account, please follow the [AWS instructions for creating an account](https://aws.amazon.com/premiumsupport/knowledge-center/create-and-activate-aws-account/). You will need a credit card to set up the account. New accounts could take up to 24 hours to be activated.
+          If you need to create an account, please follow the [AWS instructions for creating an account](https://aws.amazon.com/premiumsupport/knowledge-center/create-and-activate-aws-account/){:target="_blank"}. You will need a credit card to set up the account. New accounts could take up to 24 hours to be activated.
 
 
 ## Step 2: Select region
@@ -31,7 +31,7 @@ Follow along with these steps and/or watch our [walk-through tutorial](./introto
 
 !!! note "AWS Region"
 
-    The default region is automatically displayed in the AWS Dashboard. The [choice of region](https://docs.aws.amazon.com/emr/latest/ManagementGuide/emr-plan-region.html) has implications on fees, speed, and performance.
+    The default region is automatically displayed in the AWS Dashboard. The [choice of region](https://docs.aws.amazon.com/emr/latest/ManagementGuide/emr-plan-region.html){:target="_blank"} has implications on fees, speed, and performance.
 
 ## Step 3: Choose virtual machine
 
@@ -45,7 +45,7 @@ Follow along with these steps and/or watch our [walk-through tutorial](./introto
 
 !!! Note "Amazon EC2"
 
-         [Amazon Elastic Cloud Computing (Amazon EC2)](https://aws.amazon.com/ec2/?ec2-whats-new.sort-by=item.additionalFields.postDateTime&ec2-whats-new.sort-order=desc) features virtual computing environments called instances. They have varying combinations of CPU, memory, storage, and networking capacity, and give you the flexibility to choose the appropriate mix of resources for your applications.  
+         [Amazon Elastic Cloud Computing (Amazon EC2)](https://aws.amazon.com/ec2/?ec2-whats-new.sort-by=item.additionalFields.postDateTime&ec2-whats-new.sort-order=desc){:target="_blank"} features virtual computing environments called instances. They have varying combinations of CPU, memory, storage, and networking capacity, and give you the flexibility to choose the appropriate mix of resources for your applications.  
 
 * Select <span class="highlight_txt">EC2 Dashboard</span> and click on <span class="highlight_txt">Launch Instance</span>:
 
@@ -75,7 +75,7 @@ An Amazon Machine Image provides the template for the root volume of an instance
 
 ## Step 4: Choose an instance type
 
-Amazon EC2 provides a wide selection of instance types optimized to fit different use cases. You can consider instances to be similar to the hardware that will run your OS and applications. [Learn more about instance types and how they can meet your computing needs](https://aws.amazon.com/ec2/instance-types/).
[…]

[…]
