### 2021-11-30 da15287
Updates to AWS tutorial  (#478)
* Updates to AWS tutorial with new screenshots and correction to AMI choice

* Fixed links and corrections to build statements

* Fixed links to open in new tab
