### 2021-12-01 #130 @allisonhorst
@labarba can you advise? do I need to `whedon set doi as archive` first?
