### 2021-12-02 #130 @labarba
Ah, yes. The final steps pre-publication are:

- author should make a tagged release of their repository, and report the version number
- editor runs `@whedon set <v##> as version`
- author should make an archive deposit of their repository on Zenodo or similar service, and reports the archive DOI here
- editor runs `@whedon set <doi> as archive`
