+++ docs/General-Tools/UNIX/index.md
@@ -1,0 +1,30 @@
+---
+layout: page
+title: Overview
+hide:
+  - toc
+---
+
+An Introduction to UNIX for Cloud Computing
+============================================
+
+This tutorial introduces the UNIX command line. It is designed for scientists and clinicians who need to use cloud-based and remote computers for basic and biomedical research.
+
+!!! note Learning Objectives
+
+	The objectives of this tutorial are to:
+
+	* Understand basic UNIX command structure 
+	* Navigate through hierarchical directory structures
+	* Read, write, create, copy, move, and remove files
+	* Understand wildcard expression 
+	* Combine commands into workflows
+
+
+Lesson name | Description
+--- | ---
+[The shell and terminal](./unix_0.md) | What is UNIX? <br />Why should I use it?
+[Navigating directories](./unix_1.md) | Where am I? <br /> What files are here?
+[Working with files](./unix_2.md) | How do I open files? <br /> How do I rename or copy files?
+[Building pipelines](./unix_3.md) | How do I combine commands to build workflows?  
+[Wrap up](./unix_4.md) | How do I save my results? <br /> What commands did I type?
