### 2021-12-06 868aec3
example lesson with side by side examples
