+++ docs/General-Tools/UNIX/unix_1.md
@@ -1,1 +1,30 @@
-# Navigating directories
+# The shell and terminal
+
+The **shell** is a computer program that uses a command line interface (CLI) to give commands made by your keyboard to your operating system. Most people are used to interacting with a graphic user interface (GUI), where you can use a combination of your mouse and keyboard to carry out commands on your computer. 
+
+We can use the shell through a **terminal** program. From the terminal. We can open programs, run analyses, create documents, delete files and create folders. 
+
+For this remote workshop, we will be using a custom created create custom computing environment using [Binder](https://mybinder.org/). Click the **launch binder** button below, wait for it to launch, then open a new terminal window by clicking **Terminal**. 
+
+[![Binder](https://binder.pangeo.io/badge_logo.svg)](https://binder.pangeo.io/v2/gh/ngs-docs/2021-remote-computing-binder/stable?urlpath=rstudio)
+
+After the server launches and you select terminal, you should see an RStudio environment and the UNIX **prompt** `$`.
+
+![](https://i.imgur.com/c3sQ3K4.png)
+
+
+
+
+!!! warning Getting help
+	Getting help
+
+	When you open up the terminal in Binder, you may see a line of text or a **prompt statement** that tells us useful things such as the name of the directory we are currently in, our username, or what computer we are currently running terminal on. Let your instructors know if you see any warning or error messages. 
+
+
+
+!!! success
+	Key points about the UNIX shell and terminal
+
+	* A shell is a program that reads commands and runs programs.
+	* We are using a remote terminal provided by myBinder.org.
+
