+++ docs/General-Tools/UNIX/unix_0.md
@@ -23,11 +23,11 @@ This tutorial introduces the UNIX command line. It is designed for scientists an
 
 Lesson name | Description
 --- | ---
-[The shell and terminal](./unix_0.md) | What is UNIX? <br />Why should I use it?
-[Navigating directories](./unix_1.md) | Where am I? <br /> What files are here?
-[Working with files](./unix_2.md) | How do I open files? <br /> How do I rename or copy files? <br /> Are my files good quality?
-[Building pipelines](./unix_3.md) | How do I combine commands to build workflows?  
-[Wrap up](./unix_4.md) | How do I save my results? <br /> What commands did I type?
+[The shell and terminal](./unix_1.md) | What is UNIX? <br />Why should I use it?
+[Navigating directories](./unix_2.md) | Where am I? <br /> What files are here?
+[Working with files](./unix_3.md) | How do I open files? <br /> How do I rename or copy files? <br /> Are my files good quality?
+[Building pipelines](./unix_4.md) | How do I combine commands to build workflows?  
+[Wrap up](./unix_5.md) | How do I save my results? <br /> What commands did I type?
 
 
 The lesson materials were adapted from the UC Davis Data Lab's [Intro to Cloud Computing](https://ngs-docs.github.io/2021-august-remote-computing/) workshop, Data Carpentry's [Introduction to the Command Line for Genomics](https://datacarpentry.org/shell-genomics/) lesson, and the Lab for Data Intensive Biology's [Advanced Beginner/Intermediate Shell](https://dib-training.readthedocs.io/en/pub/2016-01-13-adv-beg-shell.html) workshop. Today's workshop is designed to prepare you for taking subsequent cloud computing and bioinformatic workshops. 
