### 2021-12-06 d082f68
add sept oct nov workshops (#474)
* add sept and oct workshops

* add november workshops

* fix conda workshop link

* hyperlink calendar text not past workshop text

* add sep oct nov workshops

* consistent titles

* added august workshops and minor edits

* updated links and minor text fixes

* update comment pr workflow

Co-authored-by: Saranya Canchi <36385305+s-canchi@users.noreply.github.com>
