+++ .github/workflows/comment_pr.yml
@@ -14,7 +14,7 @@ jobs:
         uses: actions/checkout@v1
 
       - name: Comment PR
-        uses: thollander/actions-comment-pull-request@master
+        uses: thollander/actions-comment-pull-request@v1
         with:
           message: '**Reviewer Checklist**<br/>
           - [ ] spelling and grammar<br/>
