### 2021-12-06 2574b3d
updated with august and july workshops
