### 2021-12-06 bfa90b4
added training mailing list
