### 2021-12-06 344c22e
Update README.md (#480)
* Update README.md

* updated with august and july workshops

* added training mailing list

* fixed email link

* minor text edit

* updated email alias

Co-authored-by: Saranya Canchi <36385305+s-canchi@users.noreply.github.com>
