+++ README.md
@@ -1,6 +1,6 @@
 # CFDE Training Resources
 
-This repository builds the public training materials for the CFDE: https://training.nih-cfde.org
+The `stable` branch of this repository builds the public training materials for the CFDE at https://training.nih-cfde.org. The `dev` branch contains updates that will be merged as part of the next [release](https://github.com/nih-cfde/training-and-engagement/blob/dev/docs/TrainingRepoReleasePlan/TrainingRepo-Release-Plan.md).
 
 The training website hosts tutorials and resources to learn how to perform bioinformatic analyses and access data. The GitHub repository contains CFDE internal resources, including the [website style guide](./docs/Website-Style-Guide/0index.md) and [website release protocol](./docs/TrainingRepoReleasePlan/TrainingRepo-Release-Plan.md).
 
@@ -13,27 +13,74 @@ We welcome suggestions for improvements as well as tutorial topics. Submit all t
 ## Workshops
 
 <div>
- <p><div style="font-size:22px; font-style: italic">Check out our events calendar for <a href="https://www.nih-cfde.org/events/" target="_blank" rel="noopener" style="color:#336699;">upcoming workshops !</a></div></p>
+ <p><div style="font-size:24px; font-style: italic"><i>Check out our <a href="https://www.nih-cfde.org/events/" target="_blank" rel="noopener" style="color:#336699;">events calendar</a> for upcoming workshops!</i></div></p>
 </div>
 <div>
- <p><div style="font-size:22px; font-style: italic">Past workshops:</a></div></p>
+ <p><div style="font-size:24px; font-style: italic">Join our <a href="mailto:CFDE-Training%20<CFDETraining+subscribe@CFDEpublic.groups.io" target="_blank" rel="noopener" style="color:#336699;">training mailing list</a> for all the latest updates.</div></p>
 </div>
 <div>
- <p><div style="font-size:19px"><a href="https://github.com/nih-cfde/training-and-engagement/wiki/An-Introduction-to-UNIX:-November-17,-2021" target="_blank" rel="noopener" style="color:#336699;">Intro to UNIX</a></div><i>Nov 17,&nbsp;2021 &nbsp;&bull;&nbsp; Virtual Workshop</i><br/></p>
+ <p><div style="font-size:24px; font-style: italic">Past workshops:</a></div></p>
 </div>
 <div>
- <p><div style="font-size:19px"><a href="https://github.com/nih-cfde/training-and-engagement/wiki/A-Hands-On-Introduction-to-AWS:-November-10,-2021" target="_blank" rel="noopener" style="color:#336699;">Intro to AWS</a></div><i>Nov 10,&nbsp;2021 &nbsp;&bull;&nbsp; Virtual Workshop</i><br/></p>
+ <p><div style="font-size:19px"><a href="https://github.com/nih-cfde/training-and-engagement/wiki/An-Introduction-to-UNIX:-November-17,-2021" target="_blank" rel="noopener" style="color:#336699;">Introduction to UNIX Command Line</a></div><i>Novemver 17,&nbsp;2021 &nbsp;&bull;&nbsp; Virtual Workshop</i><br/></p>
 </div>
 <div>
[…]
