### 2021-12-07 afc8cb4
added release notes for december (#482)
