+++ docs/Release-Notes/June-2021.md
@@ -1,0 +1,23 @@
+---
+layout: page
+title: June 2021 Release
+---
+
+Latest Release
+=================
+
+**Updated June 25, 2021**
+
+<span style="font-size:24px;">New Tutorials
+
+- [Introduction to Terra](../Cloud-Platforms/Introduction-to-Terra/index.md)
+- [Introduction to Conda](../General-Tools/Introduction-to-Conda/index.md)
+
+<span style="font-size:24px;">Updates and Fixes
+
+- [Snakemake](../General-Tools/Snakemake/index.md): Updated code for samtools import and added notes on generating workflow DAG
+- Updated past workshops list and notes links
+
+<span style="font-size:24px;">Website Features
+
+- Navbar titles: Reorganized the lessons into new, topic-based navbar titles
