### 2021-12-09 33109cf
Fixing conflicts in dev (#487)
* adding analytics to header.html (#353)

* remove analytics from main.html

* add matomo block to home.html

* remove from home

* add old code block setID3

* add analytics block - overrides/main.html

* full options matomo code

* remove block from main.html

* add block to home.html

* remove analytics block

* try super block header

* remove matomo again

* try matomo in header.html again

* remove spaces

* remove spaces at end

* remove '?' from lesson outline

* Updates to home page and RNA-Seq on Cavatica lesson (#470)

* Update rna_seq_2.md

Made some 
[…]
