### 2021-12-10 99d26dc
December 2021 Release (#485)
* Update rna_seq_2.md

Made some small changes, adjusted wording to be more clear.

* Update rna_seq_3.md

* Update rna_seq_4.md

* Update rna_seq_5.md

* Update rna_seq_6.md

* Update rna_seq_7.md

* Update rna_seq_8.md

* Update rna_seq_1.md

* minor text edits for conda lesson

* Update rna_seq_2.md

* Update rna_seq_3.md

* Update rna_seq_3.md

* Update rna_seq_4.md

* Update rna_seq_4.md

* Adding workshop resources  (#469)

* adding conda workshop screenshots for attendees

* communication image

* resize communication image

* adding launch binder and export
[…]
