+++ .github/actions/auto_assign_stable.yml
@@ -7,10 +7,10 @@ addAssignees: false
 # A list of reviewers to be added to pull requests (GitHub user name)
 reviewers:
   - s-canchi
-  - marisalim
-  - abhijna
+  - raynamharris
+  - ACharbonneau
+  - ctb
   - jeremywalter
-  - josesanchez1815
 
 # A number of reviewers added to the pull request
 # Set 0 to add all the reviewers (default: 0)
