### 2021-12-10 #130 @marisalim
Hi @allisonhorst!
- the GitHub tag is: https://github.com/nih-cfde/training-and-engagement/releases/tag/2021.12
- this version of the repo is archived on Zenodo here: https://zenodo.org/record/5773212#.YbPeOVmIbb0
