### 2021-12-12 96aafb4
Create spellcheck.yml
