+++ .github/workflows/spellcheck.yml
@@ -5,12 +5,26 @@ on:
     types: [opened, reopened]
 
 jobs:
-  build:
-    name: Spellcheck
+  spelling:
+    name: Spell checking
     runs-on: ubuntu-latest
     steps:
-    # The checkout step
-    - uses: actions/checkout@v1
-    - uses: rojopolis/spellcheck-github-actions@0.18.0
-      name: Spellcheck
-      
+    - name: checkout-merge
+      if: "contains(github.event_name, 'pull_request')"
+      uses: actions/checkout@v2
+      with:
+        ref: refs/pull/${{github.event.pull_request.number}}/merge
+    - name: checkout
+      if: ${{ github.event_name == 'push' ||
+        (
+          contains(github.event.comment.body, '@check-spelling-bot apply')
+        ) }}
+      uses: actions/checkout@v2
+    - uses: check-spelling/check-spelling@main
+      id: spelling
+      if: ${{ github.event_name != 'issue_comment' ||
+        (
+          contains(github.event.comment.body, '@check-spelling-bot apply')
+        ) }}
+      with:
+        experimental_apply_changes_via_bot: 1
