### 2021-12-13 d8f543b
Fixing typo issues on home page  (#493)
* Update comment_pr.yml

* Update README.md

* Update home.html

Co-authored-by: Rayna M Harris <raynamharris@users.noreply.github.com>
