+++ .github/workflows/comment_pr.yml
@@ -16,12 +16,13 @@ jobs:
       - name: Comment PR
         uses: thollander/actions-comment-pull-request@v1
         with:
-          message: '**Reviewer Checklist**<br/>
-          - [ ] spelling and grammar<br/>
-          - [ ] successful run of all installation and code chunks<br/>
-          - [ ] sufficient explanation and details for the tutorial content<br/>
-          - [ ] code syntax and naming convention<br/>
-          - [ ] adherence of tutorial format to style guide<br/>
-          - [ ] functional links (inter and intra)<br/>
-          - [ ] accessible hyperlink text'
+          message: |
+            ## Reviewer Checklist 
+            - [ ] spelling and grammar
+            - [ ] successful run of all installation and code chunks
+            - [ ] sufficient explanation and details for the tutorial content
+            - [ ] code syntax and naming convention
+            - [ ] adherence of tutorial format to style guide
+            - [ ] functional links (inter and intra)
+            - [ ] accessible hyperlink text
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
