+++ hackmd/UNIX.md
@@ -1,12 +1,16 @@
-# An Introduction to UNIX  
+# An Introduction to UNIX for Cloud Computing - December 8, 2021
 
-**When:**  
+**When:** Wednesday, December 8th from 10:00 am-12:00 pm PDT
 
-**Instructors:** 
+**Instructors:** Dr. Rayna Harris
 
-**Helpers:**   
+**Helpers:**  Dr. Saranya Canchi and Jeremy Walter
 
-**Zoom:**  
+**Zoom:** https://zoom.us/j/7575820324?pwd=d2UyMEhYZGNiV3kyUFpUL1EwQmthQT09
+
+**Slides:** https://osf.io/dw3he/
+
+**Computing environment:** [![Binder](https://binder.pangeo.io/badge_logo.svg)](https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio)
 
 **Cheatsheet:** https://training.nih-cfde.org/en/latest/General-Tools/Cheat-Sheets/bash_cheatsheet/
 
@@ -24,6 +28,8 @@ This free 2-hour workshop introduces the UNIX command line. It was designed for 
 * Understand wildcard expression 
 * Combine commands into workflows
 :::
+
+
 
 The lesson materials were adapted from the UC Davis Data Lab's [Intro to Cloud Computing](https://ngs-docs.github.io/2021-august-remote-computing/) workshop, Data Carpentry's [Introduction to the Command Line for Genomics](https://datacarpentry.org/shell-genomics/) lesson, and the Lab for Data Intensive Biology's [Advanced Beginner/Intermediate Shell](https://dib-training.readthedocs.io/en/pub/2016-01-13-adv-beg-shell.html) workshop. Today's workshop is designed to prepare you for taking subsequent cloud computing and bioinformatic workshops. 
 
@@ -45,28 +51,17 @@ While we wait to get started --
 [TOC]
 
 
-
-
-
-
-
-
-
 ---
 
 ## 1. The shell and terminal
 
 The **shell** is a computer program that uses a command line interface (CLI) to give commands made by your keyboard to your operating system. Most people are used to interacting with a graphic user interface (GUI), where you can use a combination of your mouse and keyboard to carry out commands on your computer. 
 
-
 We can use the shell through a **terminal** program. From the terminal. We can open programs, run analyses, create documents, delete files and create folders. 
 
-
-
- 
 For this remote workshop, we will be using a custom created create custom computing environment using [Binder](https://mybinder.org/). Click the **launch binder** button below, wait for it to launch, then open a new terminal window by clicking **Terminal**. 
 
-[![Binder](https://binder.pangeo.io/badge_logo.svg)](https://binder.pangeo.io/v2/gh/ngs-docs/2021-remote-computing-binder/stable?urlpath=rstudio)
+[![Binder](https://binder.pangeo.io/badge_logo.svg)](https://mybinder.org/v2/gh/nih-cfde/training-rstudio-binder/data?urlpath=rstudio)
 
 After the server launches and you select terminal, you should see an RStudio environment and the UNIX **prompt** `$`.
 
@@ -76,8 +71,6 @@ React with a :heavy_check_mark: when you have a terminal window open.
 React with a :negative_squared_cross_mark: if you have issues opening a terminal.
 
 
-
-
 :::danger
 #### Getting help
 
@@ -85,13 +78,11 @@ When you open up the terminal in Binder, you may see a line of text or a **promp
 
 :::
 
-
-
 :::success
 #### Key points about the UNIX shell and terminal
 
 * A shell is a program that reads commands and runs programs.
[…]
