+++ hackmd/UNIX.md
@@ -1,1 +1,796 @@
+# An Introduction to UNIX - November 17, 2021
+
+**When:** Wednesday, November, 17th from 10:00 am-12:00 pm PDT
+
+**Instructors:** Dr. Rayna Harris
+
+**Helpers:**  Dr. Saranya Canchi and Jeremy Walter
+
+**Zoom:** https://zoom.us/j/7575820324?pwd=d2UyMEhYZGNiV3kyUFpUL1EwQmthQT09
+
+**Slides:** https://osf.io/tvnjq/
+
+**Cheatsheet:** https://training.nih-cfde.org/en/latest/General-Tools/Cheat-Sheets/bash_cheatsheet/
+
+
+#### Workshop Overview  
+
+This free 2-hour workshop introduces the UNIX command line. It was designed for scientists and clinicians who need to use cloud-based and remote computers for basic and biomedical research. 
+
+:::success
+#### Today's Learning Goals
+
+* Understand basic UNIX command structure 
+* Navigate through hierarchical directory structures
+* Read, write, create, copy, move, and remove files
+* Understand wildcard expression 
+* Combine commands into workflows
+:::
+
+
+
+The lesson materials were adapted from the UC Davis Data Lab's [Intro to Cloud Computing](https://ngs-docs.github.io/2021-august-remote-computing/) workshop, Data Carpentry's [Introduction to the Command Line for Genomics](https://datacarpentry.org/shell-genomics/) lesson, and the Lab for Data Intensive Biology's [Advanced Beginner/Intermediate Shell](https://dib-training.readthedocs.io/en/pub/2016-01-13-adv-beg-shell.html) workshop. Today's workshop is designed to prepare you for taking subsequent cloud computing and bioinformatic workshops. 
+
+
+:::info
+While we wait to get started --
+
+1. :heavy_check_mark: Have you looked at the [pre-workshop resources page](https://github.com/nih-cfde/training-and-engagement/wiki/Resources-for-Workshop-Attendees)?
+
+2. :pencil: Please fill out our pre-workshop survey if you have not already done so:
+[Click here!](https://forms.gle/kf87XXWa7cZFDXUA9)
+
+
+3. :hand: Raise your hand on Zoom when you've completed the pre-workshop survey.
+:::
+
+
+#### Table of Contents
+[TOC]
+
+
+
+
+
+
+
+
+
+---
+
+## 1. The shell and terminal
+
+The **shell** is a computer program that uses a command line interface (CLI) to give commands made by your keyboard to your operating system. Most people are used to interacting with a graphic user interface (GUI), where you can use a combination of your mouse and keyboard to carry out commands on your computer. 
+
+
+We can use the shell through a **terminal** program. From the terminal. We can open programs, run analyses, create documents, delete files and create folders. 
+
+
+
+ 
+For this remote workshop, we will be using a custom created create custom computing environment using [Binder](https://mybinder.org/). Click the **launch binder** button below, wait for it to launch, then open a new terminal window by clicking **Terminal**. 
+
+[![Binder](https://binder.pangeo.io/badge_logo.svg)](https://binder.pangeo.io/v2/gh/ngs-docs/2021-remote-computing-binder/stable?urlpath=rstudio)
+
+After the server launches and you select terminal, you should see an RStudio environment and the UNIX **prompt** `$`.
+
+![](https://i.imgur.com/c3sQ3K4.png)
+
+React with a :heavy_check_mark: when you have a terminal window open.
+React with a :negative_squared_cross_mark: if you have issues opening a terminal.
+
+
+
+
+:::danger
+#### Getting help
+
[…]
