### 2021-12-27 #130 @labarba
Congratulations, @marisalim, your JOSE paper is accepted! 🚀

Huge thanks to our Editor: @allisonhorst, and the Reviewers: @colbrydi, @shanamatthews — we couldn't do this without you 🙏
